test: repair failing tests
